feature: hero connect with backend
